+++ fullsimpe/SimPe/SimPe/Hidden.cs
@@ -194,7 +194,7 @@ namespace SimPe
 				8.25F,
 				System.Drawing.FontStyle.Regular,
 				System.Drawing.GraphicsUnit.Point,
-				((byte)(0))
+				0
 			);
 			FormBorderStyle = FormBorderStyle.FixedDialog;
 			Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
