+++ fullsimpe/SimPe/SimPe/PackedFiles/Wrapper/SimBusinessList.cs
@@ -80,7 +80,7 @@ namespace SimPe.PackedFiles.Wrapper
 				8.25F,
 				System.Drawing.FontStyle.Regular,
 				System.Drawing.GraphicsUnit.Point,
-				((byte)(0))
+				0
 			);
 			Name = "SimBusinessList";
 			Size = new System.Drawing.Size(150, 24);
